requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.2
sqlalchemy==1.4.25
sshtunnel==0.4.0
